Switch __init__ to import
